# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Model API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/intelligence/metric-definitions": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Metric Definitions",
        "operationId": "metric_definitions_api_v1_intelligence_metric_definitions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetricDefinition"
                  },
                  "title": "Response Metric Definitions Api V1 Intelligence Metric Definitions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls/{call_id}/process": {
      "post": {
        "tags": [
          "intelligence"
        ],
        "summary": "Process Intelligence Call",
        "operationId": "process_intelligence_call_api_v1_intelligence_calls__call_id__process_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallIntelligenceDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Search Calls",
        "operationId": "search_calls_api_v1_intelligence_calls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          },
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 30
                },
                {
                  "type": "null"
                }
              ],
              "title": "Review Status"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntelligenceSearchResult"
                  },
                  "title": "Response Search Calls Api V1 Intelligence Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls/{call_id}": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Intelligence Call",
        "operationId": "intelligence_call_api_v1_intelligence_calls__call_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallIntelligenceDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls/{call_id}/transcript-corrections": {
      "post": {
        "tags": [
          "intelligence"
        ],
        "summary": "Correct Transcript",
        "operationId": "correct_transcript_api_v1_intelligence_calls__call_id__transcript_corrections_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptCorrection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallIntelligenceDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls/{call_id}/review": {
      "post": {
        "tags": [
          "intelligence"
        ],
        "summary": "Review Intelligence",
        "operationId": "review_intelligence_api_v1_intelligence_calls__call_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligenceReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/calls/{call_id}/reconcile": {
      "post": {
        "tags": [
          "intelligence"
        ],
        "summary": "Reconcile Call",
        "operationId": "reconcile_call_api_v1_intelligence_calls__call_id__reconcile_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/intelligence/dashboard": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v1_intelligence_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceDashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CallIntelligenceDetail": {
        "properties": {
          "intelligence": {
            "$ref": "#/components/schemas/IntelligenceRead"
          },
          "transcript": {
            "$ref": "#/components/schemas/TranscriptRevisionRead"
          },
          "fact": {
            "$ref": "#/components/schemas/CallFactRead"
          },
          "pipeline": {
            "$ref": "#/components/schemas/PipelineRunRead"
          },
          "reconciliations": {
            "items": {
              "$ref": "#/components/schemas/ReconciliationRead"
            },
            "type": "array",
            "title": "Reconciliations"
          }
        },
        "type": "object",
        "required": [
          "intelligence",
          "transcript",
          "fact",
          "pipeline",
          "reconciliations"
        ],
        "title": "CallIntelligenceDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntelligenceReview": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "accept",
              "correct",
              "reject"
            ],
            "title": "Decision"
          },
          "corrected_fields": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IntelligenceCorrections"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "IntelligenceReview"
      },
      "IntelligenceSearchResult": {
        "properties": {
          "call_id": {
            "type": "string",
            "title": "Call Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "confidence_bps": {
            "type": "integer",
            "title": "Confidence Bps"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "call_id",
          "topic",
          "intent",
          "summary",
          "outcome",
          "confidence_bps",
          "status",
          "started_at",
          "snippet"
        ],
        "title": "IntelligenceSearchResult"
      },
...
      "MetricDefinition": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "definition",
          "unit"
        ],
        "title": "MetricDefinition"
      },
...
      "ReconciliationCreate": {
        "properties": {
          "provider_snapshot_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Provider Snapshot Id"
          },
          "provider_status": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Provider Status"
          },
          "provider_duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Provider Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "provider_snapshot_id",
          "provider_status",
          "provider_duration_ms"
        ],
        "title": "ReconciliationCreate"
      },
...
      "TranscriptCorrection": {
        "properties": {
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranscriptSegmentCorrection"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Segments"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "segments",
          "reason"
        ],
        "title": "TranscriptCorrection"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}